Core: Fix compile (Thanks to Sovak for reporting). Remove many warnings and Fix possible out of bounds access to Void storage

+++ src/server/scripts/Kalimdor/RuinsOfAhnQiraj/boss_moam.cpp
@@ -112,7 +112,7 @@ class boss_moam : public CreatureScript
 
                 events.Update(diff);
 
-                if (me->GetPower(POWER_MANA) == me->GetMaxPower(POWER_MANA))
+                if (uint32(me->GetPower(POWER_MANA)) == me->GetMaxPower(POWER_MANA))
                 {
                     if (_isStonePhase)
                         DoAction(ACTION_STONE_PHASE_END);
